Fix decimal precision loss for percentage adjustments
Changed adjustment value storage from integer to basis points (hundredths
of percent) to preserve decimal precision for percentage adjustments like
-49.19%.

Storage strategy:
- PERCENT: stored in basis points (multiply by 100)
  Example: -49.19% → -4919 basis points
- FIXED_NET/FIXED_GROSS/SET_NET/SET_GROSS: stored in cents (unchanged)

Changes made:
1. DTOs updated to accept Double values:
   - AdjustmentRequest (check-in)
   - ServiceAdjustment (visit services)
   - PriceAdjustmentRequest (appointments)
   - AdjustmentResponse (visit details)
   - PriceAdjustmentInfo (appointment list)

2. Handlers updated to convert PERCENT to basis points:
   - CreateVisitFromReservationHandler
   - SaveVisitServicesHandler
   - CreateAppointmentHandler
   - UpdateAppointmentHandler

3. Price calculators updated for basis points:
   - AppointmentLineItem.calculateFinalNet()
   - PriceCalculator.calculateFinalNet()
   Changed division from 100 to 10000 for PERCENT type

4. Response mappers updated to convert back to percentage:
   - VisitController.mapToServiceLineItemResponse()
   - ListAppointmentsHandler service mapping

This preserves exact decimal values throughout the system:
- Frontend sends -49.19
- Backend stores as -4919 (basis points)
- Calculations use -4919 / 10000 = -0.4919
- API returns -49.19 to frontend

+++ src/main/kotlin/pl/detailing/crm/appointment/create/CreateAppointmentRequest.kt
@@ -77,7 +77,7 @@ data class ServiceLineItemRequest(
 
 data class PriceAdjustmentRequest(
     val type: AdjustmentType,
-    val value: Long
+    val value: Double  // Double to support decimal percentages like -49.19
 )
 
 data class ScheduleRequest(
